Assignment 2: Polymorphism
Design a Java interface called Lockable that includes the following methods: setKey, lock, unlock, and locked. The setKey, lock, and unlock methods take an integer parameter that represent the key. The setKey method establishes the key. The lock and unlock methods lock and unlock the object, but only if the key passed in is correct. The locked method returns a boolean that indicates whether or not the object is locked. A Lockable object represents an object whose regular methods are protected. If the object is locked, then the methods cannot be invoked; if it is unlocked, they can be invoked. Now, redesign the Coin class from Chapter 5 in your course textbook so that they are Lockable. Create a driver class to demonstrate the behaviours of Lockable objects using Coin classes.