Assignment 4: Introduction to Collections-Stacks
Create a generic class with a type parameter that simulates drawing an item at random out of a box. For example, the box might contain Strings representing names written on slips of paper, or the box might contain Integers representing the possible numbers for a lottery draw. Include the following methods:
    add() – that allows the user to add an object of the specified type to the box
    isEmpty() – to determine if the box is empty
    drawItem() – to randomly select an object from the box and return it, as well as eliminating it from the box (if the box is empty return null)
    toString() – to output the current contents of the box.
    In the driver, create two boxes, one with the names of five people and another with numbers between one and five that represent seating at a table. Use the drawItem() method to determine which seat a person will occupy.